refactor: boardDetail, postBoard refactor
